pyproject.toml
[project]
name = "llm-colosseum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dacite>=1.8.1",
    "diambra>=0.0.19",
    "diambra-arena>=2.2.7",
    "diambra-engine>=2.2.4",
    "farama-notifications>=0.0.4",
    "grpcio>=1.67.0",
    "gymnasium>=1.0.0",
    "llama-index>=0.11.19",
    "llama-index-llms-anthropic>=0.2.1",
    "llama-index-llms-bedrock>=0.2.2",
    "llama-index-llms-cerebras>=0.1.0",
    "llama-index-llms-groq>=0.2.0",
    "llama-index-llms-mistralai>=0.2.7",
    "llama-index-llms-ollama>=0.3.4",
    "llama-index-llms-openai>=0.2.15",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "numpy>=1.26.4",
    "openai>=1.52.0",
    "opencv-python>=4.10.0.84",
    "pandas==2.2.1",
    "phospho>=0.3.40",
    "pydantic>=2.9.2",
    "pyparsing>=3.2.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.2",
    "tqdm>=4.66.5",
    "llama-index-llms-langchain>=0.4.2",
    "langchain-openai>=0.2.3",
    "langchain-community>=0.3.3",
    "fastapi>=0.115.2",
    "uvicorn>=0.32.0",
]
